+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 [1.0.0]
 
+**Fixed**
+* Fixed a client-side crash when rendering the Fireplace Cornice 
+
+**Added**
+* NeoForge & Fabric config to remove the Fog Effect in SnowyBiomes.
+
+***
+
+[1.0.0]
+
 **Initial Release**
 
 ***